Replace moshi with gson
This library should also build as part of the Android source tree, but
AOSP does not have moshi codegen support. Since we only use the JSON
adapters for tests, let's swap to using gson instead.

gson is admittedly slower and disrecommended for Android code, but
since we're only using it to generate and load test data, it's fine.
In fact, as a bonus, this removes the moshi dep from production bits.

PiperOrigin-RevId: 794648170