Add KeyAttestationReason to CertPathValidatorException.
This will allow callers to distinguish between different errors in the
chain without string matching. This requires explicitly stating the
required API for each method which diretly or indirectly uses
KeyAttestationReason, but shouldn't be an actual change in requirements
since the other reasons also require API 24 when used on Android.

This also moves the key attestation chain format checks to be
exclusively in the validator instead of some of them being duplicated
in the KeyAttestationCertPath constructor. The validator checks could
previously never be executed and the CertPathValidatorException allows
for more structured information returned on exception.

PiperOrigin-RevId: 805800611